Update pipeline to use cleanser 1.2
Cleanser can now handle mudata files directly so
we only need a separate script for assigning via
umi-threshold.

I've refactored so that the cleanser and umi-threshold
paths are more separate.
